requirements.txt
# LangGraph & LangChain - Stable 0.2.x versions (tested and compatible)
langgraph==0.2.28
langchain-openai==0.2.1
langchain-anthropic==0.2.3
langchain-core==0.3.15

# Google Drive API
google-api-python-client==2.147.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Document Processing
python-docx==1.1.0

# Web Framework
flask==3.0.3
flask-cors==4.0.1

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2

# Development
pytest==8.3.3
black==24.8.0
